thematicBlocks_index_update
The names of the thematic blocks were added. At the main page I've added divs with the key informations on when the registration closes, as required by the Professour, and date and place to be more, how to put it, eye catching. It still needs formating: font size and colour change and div background color.